Stopping
Ensure files are closed when applications exits on Ctrl+C.
